refactor: 💡 update naming of components
